requirements.txt
alembic==1.4.2
appdirs==1.4.3
asn1crypto==1.4.0
astroid==2.3.3
attrs==19.3.0
azure-common==1.1.25
azure-core==1.8.0
azure-storage-blob==12.3.2
bcrypt==3.1.7
beautifulsoup4==4.8.2
boto3==1.14.39
botocore==1.17.39
bs4==0.0.1
cached-property==1.5.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==2.9.2
daiquiri==2.0.0
defusedxml==0.6.0
docutils==0.15.2
flasgger==0.9.5
Flask==1.1.2
Flask-HTTPAuth==4.1.0
Flask-Login==0.5.0
flask-marshmallow==0.13.0
Flask-Migrate==2.5.3
Flask-Seeder==1.1.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
fprint==1.0
future==0.18.2
idna==2.9
importlib-metadata==1.7.0
isodate==0.6.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
jsonpickle==1.4.1
jsonschema==3.2.0
lazy-object-proxy==1.4.3
lxml==4.5.0
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.7.1
marshmallow-sqlalchemy==0.23.1
mccabe==0.6.1
mistune==0.8.4
msrest==0.6.18
netsuitesdk==1.1.0
nose==1.3.7
numpy==1.18.1
O365==2.0.10
oauthlib==3.1.0
oscrypto==1.2.1
pandas==1.0.1
progressbar==2.5
prompt-toolkit==3.0.3
pycparser==2.20
pycryptodome==3.9.7
pycryptodomex==3.9.8
pyfiglet==0.8.post1
PyJWT==1.7.1
pylint==2.4.4
pyOpenSSL==19.1.0
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
python-json-logger==0.1.11
python-tds==1.10.0
pytz==2019.3
PyYAML==5.3.1
Random-Word==1.0.4
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
s3transfer==0.3.3
selenium==3.141.0
six==1.15.0
snowflake-connector-python==2.2.10
soupsieve==2.0
SQLAlchemy==1.3.18
stringcase==1.2.0
tabulate==0.8.6
tqdm==4.43.0
twilio==6.42.0
typed-ast==1.4.1
tzlocal==2.1
unicodecsv==0.14.1
urllib3==1.25.8
wcwidth==0.1.8
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.11.2
WTForms==2.3.3
XlsxWriter==1.2.8
xmlformatter==0.1.1
xmltodict==0.12.0
zeep==3.4.0
zipp==3.1.0
